requirements.txt
Flask==3.0.3
Pillow==10.3.0
pypdf==4.1.0
